Refactor linking helpers
